\name{RFinanceYJ-package}
\alias{RFinanceYJ-package}
\alias{RFinanceYJ}
\docType{package}
\title{
Japanese stock market from Yahoo!-finance-Japan
}
\description{
Japanese stock market from Yahoo!-finance-Japan
}
\details{
\tabular{ll}{
Package: \tab RFinanceYJ\cr
Type: \tab Package\cr
Version: \tab 0.0.3\cr
Date: \tab 2010-01-23\cr
License: \tab New BSD License\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Yohei Sato <yokkuns@tkul.jp>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
